Refactor requestBody buy ability

+++ src/main/java/com/project/realestate/dto/buyability/BuyAbilityReq.java
@@ -3,14 +3,12 @@ package com.project.realestate.dto.buyability;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 public class BuyAbilityReq {
-    private BigDecimal propertyPrice;
-    private BigDecimal downPayment;
-    private BigDecimal monthlyIncome;
+    private Double propertyPrice;
+    private Double downPayment;
+    private Double monthlyIncome;
     private Integer loanTenure;
     private Integer annualInterestRate;
 }
